requirements.txt
# oTree-may-overwrite-this-file
# IF YOU MODIFY THIS FILE, remove this comment.
# otherwise, oTree will automatically overwrite it.
otree>=5.0.0a21
psycopg2>=2.8.4
PyYAML>=6.0
